feat: Adding custom route to get total delegated supply from dao; refactor: Creating a modules folder and putting all the daos indexers inside; feat: Adding lodash utility lib;
